Ensure pixels reset correctly
If a user clicked on a single pixel to color it, that pixel would not have the
"wipe" class applied to it unless there was another "mouseup" event. If
the user clicked a pixel immediately before resetting the board, that
pixel would not be reset until another mouse up event was triggered on
the board.

If we just add the wipe class immediately after changing the background
color, this would cause it to animate in. This isn't the behavior we are
looking for, so we set a .5s delay before applying the "wipe" class.